Client Refactor | working and tested
