+++ src/client/MainClient.java
@@ -1,5 +1,9 @@
 package client;
 
+import client.controller.ClientController;
+import client.model.ClientModel;
+import client.view.ClientView;
+
 public class MainClient {
     public static void main(String[] args) {
         ClientView view = new ClientView();
